requirements.txt
awscli
pypdf
streamlit
llama_index
boto3
llama-index-llms-bedrock
faiss-cpu
langchain
openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
html2text
tiktoken
selenium
readability-lxml
openpyxl 
openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
html2text
tiktoken
selenium
readability-lxml
streamlit
streamlit-tags
google-generativeai 
scraper
pymongo
